PrivateAI/requirements.txt
# Core dependencies
mitmproxy==9.0.1
transformers==4.30.2
torch
torchvision
spacy==3.5.0
Flask==2.2.2
Flask-WTF==1.0.1
Flask-BasicAuth==0.2.0
python-dotenv==0.21.0
presidio-analyzer==2.2.32
presidio-anonymizer==2.2.32
requests==2.28.1
cryptography==39.0.1

# Production server
waitress==2.1.2
gunicorn==21.2.0

# Security and performance
flask-talisman==1.0.0  # Security headers
flask-compress==1.13   # Compression for better performance
flask-caching==2.0.2   # Caching support

# Monitoring and logging
prometheus-flask-exporter==0.22.4  # Metrics for monitoring

# Language models
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl